Write tests for this feature
In short, I have written two tests, one which tests the whether `extract_message_links` is called on message edits or not. And the second one to test the regex of `extract_message_links` and assert the message link embeds sent by it.

Special thanks to kwzrd💜#1198 for helping me out with it.